split less compiler tasks by themes type. Another builder optimize.
