+++ gulp/common/worker.js
@@ -66,7 +66,8 @@ try {
    const fs = require('fs-extra'),
       workerPool = require('workerpool'),
       { compileEsAndTs } = require('../../lib/compile-es-and-ts'),
-      { buildLess } = require('../../lib/build-less'),
+      { buildLess } = require('../../lib/less/build-less'),
+      buildNewLess = require('../../lib/less/build-new-less'),
       parseJsComponent = require('../../lib/parse-js-component'),
       processingRoutes = require('../../lib/processing-routes'),
       runMinifyCss = require('../../lib/run-minify-css'),
@@ -215,6 +216,7 @@ try {
       parseJsComponent: wrapWorkerFunction(parseJsComponent),
       parseRoutes: wrapWorkerFunction(processingRoutes.parseRoutes),
       buildLess: wrapWorkerFunction(buildLess),
+      buildNewLess: wrapWorkerFunction(buildNewLess),
       compileEsAndTs: wrapWorkerFunction(compileEsAndTs),
       buildTmpl: wrapWorkerFunction(buildTmpl),
       buildHtmlTmpl: wrapWorkerFunction(buildHtmlTmpl),
